Allow configuration of groups to be skipped during installation
Current setting only skipped development and test environment, but
depending on your setup, that might not be flexible enough to
exclude groups like ci/travis or similar.

Having this as an option (bundle_withouts) gives you the freedom
to set that to other environments, like 'development:test:staging'
or 'production' if you're deploying to a CI environment for
example.