Conditional support for writing a REVISION file
As mentioned in #72 [0], having the exact sha of the deployed version of
an application available in the root of the project can be very
handy. For example: it's easy to retrieve the deployed version when the
file is there.

If the user now sets the `git_revision_file` in her `deploy.rb` config
file:

```
set :git_revision_file, 'REVISION'
```

A file named 'REVISION' will be written during deployment which contains
the sha of the project being deployed (obtained via `git rev-parse
HEAD`).

Existing projects won't noticed any difference since the
`git_revision_file` flag must be set to have any effect on the deploy.