Ability to customise Rails' public assets path
If you have a custom location of Rails in your repository, you are now
able to specify the location of the public/assets folder - for example,
'rails/public/assets'.

Without this, when the script to check for the asset differences runs
(`check_for_changes_script`), and this path is wrong, the check to see
if this path exists will fail.

Most of the rest of the variables seem to be able to be customised if
they are non-default, adding this allows for the everything to be.